RUM-17303: Moved AppStateProviders to DatadogCore.
They are only needed in Core, and this allows keeping all the logic (in macOS case, shared between ApplicationStatePublisher and AppStateProvider) in the same module.
The protocols remain in DatadogInternal.
This allowed re-writing AppStateProvider in a better way, without duplicating the low level logic of determining the app state.
