+++ DatadogInternal/Sources/Context/AppState.swift
@@ -233,91 +233,3 @@ public struct AppStateHistory: Codable, Equatable {
         }
     }
 }
-
-#if canImport(WatchKit)
-
-import WatchKit
-
-/// Default provider for WatchOS.
-///
-/// See: https://developer.apple.com/documentation/watchkit/wkapplication/applicationstate
-public struct DefaultAppStateProvider: AppStateProvider {
-    public init() {}
-
-    /// Gets the current application state.
-    public var current: AppState {
-        return AppState(WKApplication.shared().applicationState)
-    }
-}
-
-extension AppState {
-    public init(_ state: WKApplicationState) {
-        switch state {
-        case .active: self = .active
-        case .inactive: self = .inactive
-        case .background: self = .background
-        @unknown default:
-            self = .active // in case a new state is introduced, default to most expected state
-        }
-    }
-}
-
-#elseif canImport(UIKit)
-
-import UIKit
-
-/// Default app state provider for iOS.
-///
-/// See: https://developer.apple.com/documentation/uikit/uiapplication/state
-public struct DefaultAppStateProvider: AppStateProvider {
-    public init() {}
-
-    /// Gets the current application state.
-    public var current: AppState {
-        let uiKitState = UIApplication.dd.managedShared?.applicationState ?? .active // fallback to most expected state
-        return AppState(uiKitState)
-    }
-}
-
-extension AppState {
-    public init(_ state: UIApplication.State) {
-        switch state {
-        case .active: self = .active
-        case .inactive: self = .inactive
-        case .background: self = .background
-        @unknown default: self = .active // in case a new state is introduced, default to most expected state
-        }
-    }
-}
-
-#elseif canImport(AppKit)
-
-import AppKit
-
-public struct DefaultAppStateProvider: AppStateProvider {
-    public init() {}
-    public var current: AppState {
-        // Note: based on empirical evidence, when an app is launched in macOS, sometimes NSApp.isActive returns
-        // true, sometimes false, when called during the AppDelegate.applicationDidFinishLaunching method. This
-        // happens even if the conditions the app is launched in are exactly the same (for example, double-clicking
-        // it in the Finder). Do not assume this will always return true if the application is launched directly
-        // to the foreground.
-        //
-        // We cannot detect here if the system broadcasted the NSWorkspace.willSleepNotification notification
-        // to running apps and the system is about to go to sleep. It's unlikely that is the case when
-        // an application is being launched, assuming Datadog SDK is being initialized at app startup,
-        // but it's possible.
-        let loginWindowActive = NSWorkspace.shared.frontmostApplication?.bundleIdentifier == "com.apple.loginwindow"
-        if loginWindowActive {
-            return .lockScreen
-        }
-
-        if DDApplication.shared.isHidden {
-            return .hidden
-        }
-
-        return DDApplication.shared.isActive ? .active : .inactive
-    }
-}
-
-#endif
